feat(town-gen): add town signs helpers and wire into generation (new worldgen/town/signs.js; imported and used in town_gen.js)
